+++ src/main/java/com/fouribnb/coupon/presentation/dto/response/GrantCouponResponseDto.java
@@ -13,6 +13,6 @@ public class GrantCouponResponseDto {
     private UUID paymentId;
     private String couponName;
     private CouponStatus couponStatus;
-    private int discountValue;
+    private Long discountValue;
     private Boolean isUsed;
 }
